correct password list import error
Fixed error when on importing passwords it would hit an exception and never load the list. Issue was with the quoting in the input strings and erroring on bad lines. Added a fix for this in the import statement and removed the ValueError exception from the try catch to give a more accurate concept of the error.  Also recasted the password column as a string.