operation optimizations
Renamed all of the outer scope refs to df to be within the analyzer class. Changed how the filter argument was passed to the init function and changed it to only calculate masks after filtering it down in attempts to solve memory issues with very large datasets. Optimized the functions used to create masks and other operations by using apply instead of the previous method for efficiency.